backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.7.3
python-dotenv==1.0.1
joblib==1.3.2
numpy==1.26.3
apscheduler==3.10.4
requests==2.31.0
pydantic==2.5.0
